backend/foodgram/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.2.0
fonttools==4.39.4
gunicorn==20.1.0
html5lib==1.1
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pyphen==0.14.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
urllib3==2.0.2
webcolors==1.13
webencodings==0.5.1
zopfli==0.2.2
